feat: update enums and move local core to core/ folder
