+++ composio/client/enums.py
@@ -16,341 +16,341 @@ class Tag(tuple, Enum):
         return self.value[0]
 
     IMPORTANT = ("default", "important")
-    ASANA_GOALS = ("asana", "Goals")
-    ASANA_TASKS = ("asana", "Tasks")
-    ASANA_PROJECT_STATUSES = ("asana", "Project statuses")
-    ASANA_ALLOCATIONS = ("asana", "Allocations")
-    ASANA_ATTACHMENTS = ("asana", "Attachments")
     ASANA_PORTFOLIOS = ("asana", "Portfolios")
-    ASANA_WEBHOOKS = ("asana", "Webhooks")
-    ASANA_WORKSPACE_MEMBERSHIPS = ("asana", "Workspace memberships")
-    ASANA_TEAMS = ("asana", "Teams")
-    ASANA_SECTIONS = ("asana", "Sections")
-    ASANA_STATUS_UPDATES = ("asana", "Status updates")
-    ASANA_PROJECT_TEMPLATES = ("asana", "Project templates")
-    ASANA_TASK_TEMPLATES = ("asana", "Task templates")
-    ASANA_GOAL_RELATIONSHIPS = ("asana", "Goal relationships")
-    ASANA_TYPEAHEAD = ("asana", "Typeahead")
-    ASANA_JOBS = ("asana", "Jobs")
-    ASANA_RULES = ("asana", "Rules")
-    ASANA_TEAM_MEMBERSHIPS = ("asana", "Team memberships")
-    ASANA_USER_TASK_LISTS = ("asana", "User task lists")
-    ASANA_AUDIT_LOG_API = ("asana", "Audit log API")
-    ASANA_USERS = ("asana", "Users")
-    ASANA_BATCH_API = ("asana", "Batch API")
-    ASANA_ORGANIZATION_EXPORTS = ("asana", "Organization exports")
-    ASANA_PROJECT_BRIEFS = ("asana", "Project briefs")
-    ASANA_PROJECT_MEMBERSHIPS = ("asana", "Project memberships")
     ASANA_PORTFOLIO_MEMBERSHIPS = ("asana", "Portfolio memberships")
-    ASANA_WORKSPACES = ("asana", "Workspaces")
-    ASANA_MEMBERSHIPS = ("asana", "Memberships")
-    ASANA_CUSTOM_FIELDS = ("asana", "Custom fields")
-    ASANA_CUSTOM_FIELD_SETTINGS = ("asana", "Custom field settings")
     ASANA_EVENTS = ("asana", "Events")
-    ASANA_STORIES = ("asana", "Stories")
+    ASANA_RULES = ("asana", "Rules")
+    ASANA_MEMBERSHIPS = ("asana", "Memberships")
+    ASANA_PROJECT_BRIEFS = ("asana", "Project briefs")
+    ASANA_TEAMS = ("asana", "Teams")
+    ASANA_ALLOCATIONS = ("asana", "Allocations")
+    ASANA_TYPEAHEAD = ("asana", "Typeahead")
+    ASANA_TASKS = ("asana", "Tasks")
+    ASANA_WORKSPACE_MEMBERSHIPS = ("asana", "Workspace memberships")
+    ASANA_WEBHOOKS = ("asana", "Webhooks")
     ASANA_PROJECTS = ("asana", "Projects")
-    ASANA_TIME_TRACKING_ENTRIES = ("asana", "Time tracking entries")
+    ASANA_PROJECT_TEMPLATES = ("asana", "Project templates")
+    ASANA_STORIES = ("asana", "Stories")
+    ASANA_ORGANIZATION_EXPORTS = ("asana", "Organization exports")
+    ASANA_TEAM_MEMBERSHIPS = ("asana", "Team memberships")
+    ASANA_AUDIT_LOG_API = ("asana", "Audit log API")
+    ASANA_TASK_TEMPLATES = ("asana", "Task templates")
+    ASANA_GOALS = ("asana", "Goals")
+    ASANA_GOAL_RELATIONSHIPS = ("asana", "Goal relationships")
+    ASANA_BATCH_API = ("asana", "Batch API")
+    ASANA_PROJECT_MEMBERSHIPS = ("asana", "Project memberships")
     ASANA_TAGS = ("asana", "Tags")
+    ASANA_CUSTOM_FIELDS = ("asana", "Custom fields")
     ASANA_TIME_PERIODS = ("asana", "Time periods")
-    ATTIO_TASKS = ("attio", "Tasks")
-    ATTIO_COMMENTS = ("attio", "Comments")
-    ATTIO_NOTES = ("attio", "Notes")
+    ASANA_ATTACHMENTS = ("asana", "Attachments")
+    ASANA_USER_TASK_LISTS = ("asana", "User task lists")
+    ASANA_CUSTOM_FIELD_SETTINGS = ("asana", "Custom field settings")
+    ASANA_USERS = ("asana", "Users")
+    ASANA_STATUS_UPDATES = ("asana", "Status updates")
+    ASANA_TIME_TRACKING_ENTRIES = ("asana", "Time tracking entries")
+    ASANA_SECTIONS = ("asana", "Sections")
+    ASANA_JOBS = ("asana", "Jobs")
+    ASANA_PROJECT_STATUSES = ("asana", "Project statuses")
+    ASANA_WORKSPACES = ("asana", "Workspaces")
     ATTIO_ATTRIBUTES = ("attio", "Attributes")
-    ATTIO_RECORDS = ("attio", "Records")
     ATTIO_OBJECTS = ("attio", "Objects")
+    ATTIO_WEBHOOKS = ("attio", "Webhooks")
+    ATTIO_THREADS = ("attio", "Threads")
     ATTIO_META = ("attio", "Meta")
     ATTIO_WORKSPACE_MEMBERS = ("attio", "Workspace members")
-    ATTIO_THREADS = ("attio", "Threads")
+    ATTIO_NOTES = ("attio", "Notes")
+    ATTIO_COMMENTS = ("attio", "Comments")
+    ATTIO_RECORDS = ("attio", "Records")
+    ATTIO_TASKS = ("attio", "Tasks")
     ATTIO_ENTRIES = ("attio", "Entries")
-    ATTIO_WEBHOOKS = ("attio", "Webhooks")
     ATTIO_LISTS = ("attio", "Lists")
-    BREVO_TASKS = ("brevo", "Tasks")
-    BREVO_EVENT = ("brevo", "Event")
-    BREVO_TRANSACTIONAL_WHATSAPP = ("brevo", "Transactional WhatsApp")
-    BREVO_EXTERNAL_FEEDS = ("brevo", "External Feeds")
-    BREVO_DEALS = ("brevo", "Deals")
-    BREVO_PROCESS = ("brevo", "Process")
     BREVO_COMPANIES = ("brevo", "Companies")
-    BREVO_WEBHOOKS = ("brevo", "Webhooks")
-    BREVO_RESELLER = ("brevo", "Reseller")
-    BREVO_COUPONS = ("brevo", "Coupons")
-    BREVO_WHATSAPP_CAMPAIGNS = ("brevo", "WhatsApp Campaigns")
-    BREVO_USER = ("brevo", "User")
-    BREVO_TRANSACTIONAL_SMS = ("brevo", "Transactional SMS")
-    BREVO_SENDERS = ("brevo", "Senders")
-    BREVO_ACCOUNT = ("brevo", "Account")
-    BREVO_MASTER_ACCOUNT = ("brevo", "Master account")
-    BREVO_TRANSACTIONAL_EMAILS = ("brevo", "Transactional emails")
-    BREVO_NOTES = ("brevo", "Notes")
-    BREVO_FILES = ("brevo", "Files")
+    BREVO_EXTERNAL_FEEDS = ("brevo", "External Feeds")
     BREVO_INBOUND_PARSING = ("brevo", "Inbound Parsing")
-    BREVO_SMS_CAMPAIGNS = ("brevo", "SMS Campaigns")
+    BREVO_ACCOUNT = ("brevo", "Account")
     BREVO_CONVERSATIONS = ("brevo", "Conversations")
+    BREVO_RESELLER = ("brevo", "Reseller")
+    BREVO_USER = ("brevo", "User")
+    BREVO_SMS_CAMPAIGNS = ("brevo", "SMS Campaigns")
+    BREVO_NOTES = ("brevo", "Notes")
+    BREVO_TASKS = ("brevo", "Tasks")
+    BREVO_WHATSAPP_CAMPAIGNS = ("brevo", "WhatsApp Campaigns")
+    BREVO_WEBHOOKS = ("brevo", "Webhooks")
+    BREVO_TRANSACTIONAL_WHATSAPP = ("brevo", "Transactional WhatsApp")
+    BREVO_COUPONS = ("brevo", "Coupons")
     BREVO_DOMAINS = ("brevo", "Domains")
-    BREVO_ECOMMERCE = ("brevo", "Ecommerce")
+    BREVO_EVENT = ("brevo", "Event")
     BREVO_EMAIL_CAMPAIGNS = ("brevo", "Email Campaigns")
+    BREVO_TRANSACTIONAL_EMAILS = ("brevo", "Transactional emails")
+    BREVO_SENDERS = ("brevo", "Senders")
+    BREVO_TRANSACTIONAL_SMS = ("brevo", "Transactional SMS")
+    BREVO_PROCESS = ("brevo", "Process")
+    BREVO_DEALS = ("brevo", "Deals")
     BREVO_CONTACTS = ("brevo", "Contacts")
-    CLICKUP_ROLES = ("clickup", "Roles")
-    CLICKUP_COMMENTS = ("clickup", "Comments")
-    CLICKUP_GOALS = ("clickup", "Goals")
-    CLICKUP_TASKS = ("clickup", "Tasks")
-    CLICKUP_AUTHORIZATION = ("clickup", "Authorization")
-    CLICKUP_ATTACHMENTS = ("clickup", "Attachments")
-    CLICKUP_TIME_TRACKING = ("clickup", "Time Tracking")
-    CLICKUP_TASK_RELATIONSHIPS = ("clickup", "Task Relationships")
-    CLICKUP_WEBHOOKS = ("clickup", "Webhooks")
-    CLICKUP_TEAMS___USER_GROUPS = ("clickup", "Teams - User Groups")
-    CLICKUP_SHARED_HIERARCHY = ("clickup", "Shared Hierarchy")
-    CLICKUP_GUESTS = ("clickup", "Guests")
-    CLICKUP_TEAMS___WORKSPACES = ("clickup", "Teams - Workspaces")
+    BREVO_MASTER_ACCOUNT = ("brevo", "Master account")
+    BREVO_ECOMMERCE = ("brevo", "Ecommerce")
+    BREVO_FILES = ("brevo", "Files")
     CLICKUP_MEMBERS = ("clickup", "Members")
-    CLICKUP_USERS = ("clickup", "Users")
-    CLICKUP_TASK_TEMPLATES = ("clickup", "Task Templates")
+    CLICKUP_GUESTS = ("clickup", "Guests")
     CLICKUP_SPACES = ("clickup", "Spaces")
-    CLICKUP_LISTS = ("clickup", "Lists")
-    CLICKUP_CUSTOM_TASK_TYPES = ("clickup", "Custom Task Types")
     CLICKUP_FOLDERS = ("clickup", "Folders")
-    CLICKUP_VIEWS = ("clickup", "Views")
-    CLICKUP_CUSTOM_FIELDS = ("clickup", "Custom Fields")
-    CLICKUP_TASK_CHECKLISTS = ("clickup", "Task Checklists")
+    CLICKUP_TEAMS___WORKSPACES = ("clickup", "Teams - Workspaces")
     CLICKUP_TIME_TRACKING__LEGACY_ = ("clickup", "Time Tracking (Legacy)")
+    CLICKUP_ROLES = ("clickup", "Roles")
+    CLICKUP_TASK_CHECKLISTS = ("clickup", "Task Checklists")
+    CLICKUP_TASKS = ("clickup", "Tasks")
+    CLICKUP_LISTS = ("clickup", "Lists")
+    CLICKUP_CUSTOM_FIELDS = ("clickup", "Custom Fields")
+    CLICKUP_TIME_TRACKING = ("clickup", "Time Tracking")
+    CLICKUP_WEBHOOKS = ("clickup", "Webhooks")
+    CLICKUP_AUTHORIZATION = ("clickup", "Authorization")
+    CLICKUP_CUSTOM_TASK_TYPES = ("clickup", "Custom Task Types")
+    CLICKUP_GOALS = ("clickup", "Goals")
+    CLICKUP_TASK_RELATIONSHIPS = ("clickup", "Task Relationships")
     CLICKUP_TAGS = ("clickup", "Tags")
-    ELEVENLABS_VOICES = ("elevenlabs", "voices")
-    ELEVENLABS_SPEECH_TO_SPEECH = ("elevenlabs", "speech-to-speech")
-    ELEVENLABS_MODELS = ("elevenlabs", "models")
-    ELEVENLABS_SPEECH_HISTORY = ("elevenlabs", "speech-history")
-    ELEVENLABS_PRONUNCIATION_DICTIONARY = ("elevenlabs", "Pronunciation Dictionary")
-    ELEVENLABS_VOICE_GENERATION = ("elevenlabs", "voice-generation")
-    ELEVENLABS_PROJECTS = ("elevenlabs", "projects")
-    ELEVENLABS_TEXT_TO_SPEECH = ("elevenlabs", "text-to-speech")
-    ELEVENLABS_WORKSPACE = ("elevenlabs", "workspace")
-    ELEVENLABS_DUBBING = ("elevenlabs", "dubbing")
-    ELEVENLABS_AUDIO_NATIVE = ("elevenlabs", "audio-native")
+    CLICKUP_VIEWS = ("clickup", "Views")
+    CLICKUP_ATTACHMENTS = ("clickup", "Attachments")
+    CLICKUP_USERS = ("clickup", "Users")
+    CLICKUP_COMMENTS = ("clickup", "Comments")
+    CLICKUP_SHARED_HIERARCHY = ("clickup", "Shared Hierarchy")
+    CLICKUP_TASK_TEMPLATES = ("clickup", "Task Templates")
+    CLICKUP_TEAMS___USER_GROUPS = ("clickup", "Teams - User Groups")
     ELEVENLABS_SAMPLES = ("elevenlabs", "samples")
+    ELEVENLABS_SPEECH_TO_SPEECH = ("elevenlabs", "speech-to-speech")
+    ELEVENLABS_VOICE_GENERATION = ("elevenlabs", "voice-generation")
+    ELEVENLABS_TEXT_TO_SPEECH = ("elevenlabs", "text-to-speech")
+    ELEVENLABS_PROJECTS = ("elevenlabs", "projects")
+    ELEVENLABS_WORKSPACE = ("elevenlabs", "workspace")
+    ELEVENLABS_AUDIO_NATIVE = ("elevenlabs", "audio-native")
+    ELEVENLABS_SPEECH_HISTORY = ("elevenlabs", "speech-history")
     ELEVENLABS_USER = ("elevenlabs", "user")
-    FIGMA_DEV_RESOURCES = ("figma", "Dev Resources")
-    FIGMA_COMMENTS = ("figma", "Comments")
-    FIGMA_PROJECTS = ("figma", "Projects")
-    FIGMA_FILES = ("figma", "Files")
-    FIGMA_COMMENT_REACTIONS = ("figma", "Comment Reactions")
-    FIGMA_USERS = ("figma", "Users")
-    FIGMA_COMPONENTS = ("figma", "Components")
-    FIGMA_STYLES = ("figma", "Styles")
-    FIGMA_COMPONENT_SETS = ("figma", "Component Sets")
-    FIGMA_ACTIVITY_LOGS = ("figma", "Activity Logs")
-    FIGMA_VARIABLES = ("figma", "Variables")
+    ELEVENLABS_DUBBING = ("elevenlabs", "dubbing")
+    ELEVENLABS_PRONUNCIATION_DICTIONARY = ("elevenlabs", "Pronunciation Dictionary")
+    ELEVENLABS_VOICES = ("elevenlabs", "voices")
+    ELEVENLABS_MODELS = ("elevenlabs", "models")
     FIGMA_PAYMENTS = ("figma", "Payments")
+    FIGMA_ACTIVITY_LOGS = ("figma", "Activity Logs")
+    FIGMA_DEV_RESOURCES = ("figma", "Dev Resources")
+    FIGMA_VARIABLES = ("figma", "Variables")
+    FIGMA_USERS = ("figma", "Users")
+    FIGMA_COMMENTS = ("figma", "Comments")
+    FIGMA_COMMENT_REACTIONS = ("figma", "Comment Reactions")
+    FIGMA_FILES = ("figma", "Files")
+    FIGMA_STYLES = ("figma", "Styles")
+    FIGMA_COMPONENTS = ("figma", "Components")
     FIGMA_WEBHOOKS = ("figma", "Webhooks")
-    GITHUB_ORGS = ("github", "orgs")
-    GITHUB_SECURITY_ADVISORIES = ("github", "security-advisories")
-    GITHUB_REPOS = ("github", "repos")
-    GITHUB_REACTIONS = ("github", "reactions")
-    GITHUB_CODE = ("github", "code")
-    GITHUB_DEPENDENCY_GRAPH = ("github", "dependency-graph")
-    GITHUB_CODESPACES = ("github", "codespaces")
-    GITHUB_CODES_OF_CONDUCT = ("github", "codes-of-conduct")
-    GITHUB_CLASSROOM = ("github", "classroom")
-    GITHUB_INTERACTIONS = ("github", "interactions")
-    GITHUB_EMOJIS = ("github", "emojis")
-    GITHUB_PULLS = ("github", "pulls")
-    GITHUB_COPILOT = ("github", "copilot")
-    GITHUB_OIDC = ("github", "oidc")
-    GITHUB_CHECKS = ("github", "checks")
-    GITHUB_BILLING = ("github", "billing")
-    GITHUB_GISTS = ("github", "gists")
-    GITHUB_ACTIONS = ("github", "actions")
-    GITHUB_PACKAGES = ("github", "packages")
-    GITHUB_PROJECTS = ("github", "projects")
-    GITHUB_META = ("github", "meta")
-    GITHUB_SECRET_SCANNING = ("github", "secret-scanning")
-    GITHUB_ACTIVITY = ("github", "activity")
-    GITHUB_TEAMS = ("github", "teams")
-    GITHUB_IMPORTANT = ("github", "important")
+    FIGMA_COMPONENT_SETS = ("figma", "Component Sets")
+    FIGMA_PROJECTS = ("figma", "Projects")
     GITHUB_GITIGNORE = ("github", "gitignore")
+    GITHUB_CODE = ("github", "code")
+    GITHUB_GISTS = ("github", "gists")
+    GITHUB_CODESPACES = ("github", "codespaces")
     GITHUB_APPS = ("github", "apps")
-    GITHUB_RATE_LIMIT = ("github", "rate-limit")
-    GITHUB_ISSUES = ("github", "issues")
-    GITHUB_SEARCH = ("github", "search")
-    GITHUB_DEPENDABOT = ("github", "dependabot")
-    GITHUB_MIGRATIONS = ("github", "migrations")
+    GITHUB_SECURITY_ADVISORIES = ("github", "security-advisories")
+    GITHUB_DEPENDENCY_GRAPH = ("github", "dependency-graph")
+    GITHUB_META = ("github", "meta")
     GITHUB_LICENSES = ("github", "licenses")
-    GITHUB_GIT = ("github", "git")
-    GITHUB_USERS = ("github", "users")
+    GITHUB_REPOS = ("github", "repos")
+    GITHUB_ACTIONS = ("github", "actions")
+    GITHUB_PROJECTS = ("github", "projects")
+    GITHUB_REACTIONS = ("github", "reactions")
+    GITHUB_CODES_OF_CONDUCT = ("github", "codes-of-conduct")
+    GITHUB_INTERACTIONS = ("github", "interactions")
+    GITHUB_PACKAGES = ("github", "packages")
+    GITHUB_PULLS = ("github", "pulls")
+    GITHUB_RATE_LIMIT = ("github", "rate-limit")
+    GITHUB_TEAMS = ("github", "teams")
+    GITHUB_BILLING = ("github", "billing")
+    GITHUB_CLASSROOM = ("github", "classroom")
+    GITHUB_MIGRATIONS = ("github", "migrations")
     GITHUB_MARKDOWN = ("github", "markdown")
+    GITHUB_GIT = ("github", "git")
+    GITHUB_SECRET_SCANNING = ("github", "secret-scanning")
+    GITHUB_CHECKS = ("github", "checks")
+    GITHUB_USERS = ("github", "users")
+    GITHUB_IMPORTANT = ("github", "important")
+    GITHUB_DEPENDABOT = ("github", "dependabot")
+    GITHUB_ACTIVITY = ("github", "activity")
     GITHUB_CODE_SCANNING = ("github", "code-scanning")
-    LISTENNOTES_INSIGHTS_API = ("listennotes", "Insights API")
-    LISTENNOTES_PLAYLIST_API = ("listennotes", "Playlist API")
+    GITHUB_ISSUES = ("github", "issues")
+    GITHUB_OIDC = ("github", "oidc")
+    GITHUB_EMOJIS = ("github", "emojis")
+    GITHUB_ORGS = ("github", "orgs")
+    GITHUB_COPILOT = ("github", "copilot")
+    GITHUB_SEARCH = ("github", "search")
     LISTENNOTES_DIRECTORY_API = ("listennotes", "Directory API")
+    LISTENNOTES_PLAYLIST_API = ("listennotes", "Playlist API")
     LISTENNOTES_SEARCH_API = ("listennotes", "Search API")
     LISTENNOTES_PODCASTER_API = ("listennotes", "Podcaster API")
-    NASA_ORGANIZATION = ("nasa", "Organization")
+    LISTENNOTES_INSIGHTS_API = ("listennotes", "Insights API")
     NASA_PROJECT = ("nasa", "Project")
+    NASA_ORGANIZATION = ("nasa", "Organization")
     NASA_RESOURCE = ("nasa", "Resource")
-    OKTA_LOG = ("okta", "Log")
-    OKTA_THREATINSIGHT = ("okta", "ThreatInsight")
-    OKTA_AUTHORIZATIONSERVER = ("okta", "AuthorizationServer")
-    OKTA_POLICY = ("okta", "Policy")
     OKTA_USERTYPE = ("okta", "UserType")
-    OKTA_TEMPLATE = ("okta", "Template")
-    OKTA_USERFACTOR = ("okta", "UserFactor")
-    OKTA_APPLICATION = ("okta", "Application")
-    OKTA_FEATURE = ("okta", "Feature")
     OKTA_INLINEHOOK = ("okta", "InlineHook")
-    OKTA_USERSCHEMA = ("okta", "UserSchema")
-    OKTA_GROUPSCHEMA = ("okta", "GroupSchema")
-    OKTA_USER = ("okta", "User")
-    OKTA_DOMAIN = ("okta", "Domain")
-    OKTA_SESSION = ("okta", "Session")
-    OKTA_LINKEDOBJECT = ("okta", "LinkedObject")
-    OKTA_AUTHENTICATOR = ("okta", "Authenticator")
-    OKTA_ORG = ("okta", "Org")
     OKTA_GROUP = ("okta", "Group")
-    OKTA_NETWORKZONE = ("okta", "NetworkZone")
-    OKTA_IDENTITYPROVIDER = ("okta", "IdentityProvider")
-    OKTA_PROFILEMAPPING = ("okta", "ProfileMapping")
+    OKTA_LINKEDOBJECT = ("okta", "LinkedObject")
+    OKTA_USERSCHEMA = ("okta", "UserSchema")
+    OKTA_AUTHORIZATIONSERVER = ("okta", "AuthorizationServer")
+    OKTA_USERFACTOR = ("okta", "UserFactor")
     OKTA_BRAND = ("okta", "Brand")
-    OKTA_SUBSCRIPTION = ("okta", "Subscription")
+    OKTA_USER = ("okta", "User")
     OKTA_TRUSTEDORIGIN = ("okta", "TrustedOrigin")
+    OKTA_LOG = ("okta", "Log")
+    OKTA_NETWORKZONE = ("okta", "NetworkZone")
+    OKTA_GROUPSCHEMA = ("okta", "GroupSchema")
+    OKTA_THREATINSIGHT = ("okta", "ThreatInsight")
+    OKTA_FEATURE = ("okta", "Feature")
+    OKTA_ORG = ("okta", "Org")
+    OKTA_SUBSCRIPTION = ("okta", "Subscription")
+    OKTA_APPLICATION = ("okta", "Application")
+    OKTA_PROFILEMAPPING = ("okta", "ProfileMapping")
+    OKTA_SESSION = ("okta", "Session")
+    OKTA_POLICY = ("okta", "Policy")
+    OKTA_IDENTITYPROVIDER = ("okta", "IdentityProvider")
     OKTA_EVENTHOOK = ("okta", "EventHook")
-    SLACK_REMINDERS = ("slack", "reminders")
-    SLACK_PINS = ("slack", "pins")
-    SLACK_APPS_PERMISSIONS_SCOPES = ("slack", "apps.permissions.scopes")
-    SLACK_AUTH = ("slack", "auth")
-    SLACK_REACTIONS = ("slack", "reactions")
-    SLACK_STARS = ("slack", "stars")
-    SLACK_USERGROUPS_USERS = ("slack", "usergroups.users")
-    SLACK_ADMIN = ("slack", "admin")
-    SLACK_ADMIN_CONVERSATIONS_RESTRICTACCESS = ("slack", "admin.conversations.restrictAccess")
+    OKTA_AUTHENTICATOR = ("okta", "Authenticator")
+    OKTA_DOMAIN = ("okta", "Domain")
+    OKTA_TEMPLATE = ("okta", "Template")
+    SLACK_TEAM_PROFILE = ("slack", "team.profile")
+    SLACK_BOTS = ("slack", "bots")
     SLACK_OAUTH_V2 = ("slack", "oauth.v2")
-    SLACK_FILES_REMOTE = ("slack", "files.remote")
-    SLACK_CHAT_SCHEDULEDMESSAGES = ("slack", "chat.scheduledMessages")
-    SLACK_ADMIN_USERS_SESSION = ("slack", "admin.users.session")
-    SLACK_ADMIN_CONVERSATIONS = ("slack", "admin.conversations")
+    SLACK_RTM = ("slack", "rtm")
+    SLACK_ADMIN_INVITEREQUESTS_APPROVED = ("slack", "admin.inviteRequests.approved")
+    SLACK_APPS = ("slack", "apps")
+    SLACK_ADMIN_TEAMS = ("slack", "admin.teams")
+    SLACK_APPS_PERMISSIONS = ("slack", "apps.permissions")
     SLACK_ADMIN_APPS = ("slack", "admin.apps")
-    SLACK_DND = ("slack", "dnd")
-    SLACK_OAUTH = ("slack", "oauth")
-    SLACK_VIEWS = ("slack", "views")
-    SLACK_EMOJI = ("slack", "emoji")
     SLACK_ADMIN_TEAMS_SETTINGS = ("slack", "admin.teams.settings")
-    SLACK_ADMIN_APPS_APPROVED = ("slack", "admin.apps.approved")
+    SLACK_CALLS_PARTICIPANTS = ("slack", "calls.participants")
+    SLACK_STARS = ("slack", "stars")
+    SLACK_ADMIN_CONVERSATIONS_RESTRICTACCESS = ("slack", "admin.conversations.restrictAccess")
+    SLACK_WORKFLOWS = ("slack", "workflows")
+    SLACK_ADMIN_USERGROUPS = ("slack", "admin.usergroups")
+    SLACK_ADMIN_CONVERSATIONS = ("slack", "admin.conversations")
+    SLACK_ADMIN_TEAMS_OWNERS = ("slack", "admin.teams.owners")
+    SLACK_CALLS = ("slack", "calls")
+    SLACK_FILES = ("slack", "files")
+    SLACK_REACTIONS = ("slack", "reactions")
+    SLACK_APPS_PERMISSIONS_USERS = ("slack", "apps.permissions.users")
+    SLACK_ADMIN_USERS_SESSION = ("slack", "admin.users.session")
+    SLACK_APPS_PERMISSIONS_RESOURCES = ("slack", "apps.permissions.resources")
+    SLACK_DND = ("slack", "dnd")
+    SLACK_ADMIN_APPS_RESTRICTED = ("slack", "admin.apps.restricted")
+    SLACK_CONVERSATIONS = ("slack", "conversations")
+    SLACK_EMOJI = ("slack", "emoji")
     SLACK_ADMIN_INVITEREQUESTS = ("slack", "admin.inviteRequests")
+    SLACK_USERGROUPS = ("slack", "usergroups")
+    SLACK_ADMIN_APPS_APPROVED = ("slack", "admin.apps.approved")
+    SLACK_TEAM = ("slack", "team")
+    SLACK_FILES_REMOTE = ("slack", "files.remote")
+    SLACK_ADMIN_APPS_REQUESTS = ("slack", "admin.apps.requests")
+    SLACK_DIALOG = ("slack", "dialog")
+    SLACK_FILES_COMMENTS = ("slack", "files.comments")
+    SLACK_APPS_EVENT_AUTHORIZATIONS = ("slack", "apps.event.authorizations")
+    SLACK_ADMIN = ("slack", "admin")
+    SLACK_CHAT_SCHEDULEDMESSAGES = ("slack", "chat.scheduledMessages")
     SLACK_USERS = ("slack", "users")
+    SLACK_VIEWS = ("slack", "views")
+    SLACK_IMPORTANT = ("slack", "important")
+    SLACK_CHAT = ("slack", "chat")
+    SLACK_OAUTH = ("slack", "oauth")
+    SLACK_AUTH = ("slack", "auth")
+    SLACK_ADMIN_USERS = ("slack", "admin.users")
+    SLACK_ADMIN_INVITEREQUESTS_DENIED = ("slack", "admin.inviteRequests.denied")
+    SLACK_USERGROUPS_USERS = ("slack", "usergroups.users")
+    SLACK_API = ("slack", "api")
+    SLACK_REMINDERS = ("slack", "reminders")
+    SLACK_ADMIN_EMOJI = ("slack", "admin.emoji")
+    SLACK_PINS = ("slack", "pins")
+    SLACK_MIGRATION = ("slack", "migration")
     SLACK_ADMIN_CONVERSATIONS_EKM = ("slack", "admin.conversations.ekm")
     SLACK_USERS_PROFILE = ("slack", "users.profile")
-    SLACK_RTM = ("slack", "rtm")
-    SLACK_FILES_COMMENTS = ("slack", "files.comments")
-    SLACK_CONVERSATIONS = ("slack", "conversations")
-    SLACK_APPS_PERMISSIONS = ("slack", "apps.permissions")
-    SLACK_ADMIN_APPS_RESTRICTED = ("slack", "admin.apps.restricted")
-    SLACK_API = ("slack", "api")
-    SLACK_CALLS_PARTICIPANTS = ("slack", "calls.participants")
-    SLACK_TEAM_PROFILE = ("slack", "team.profile")
-    SLACK_CALLS = ("slack", "calls")
-    SLACK_ADMIN_TEAMS = ("slack", "admin.teams")
-    SLACK_ADMIN_USERGROUPS = ("slack", "admin.usergroups")
-    SLACK_IMPORTANT = ("slack", "important")
-    SLACK_APPS = ("slack", "apps")
-    SLACK_SEARCH = ("slack", "search")
-    SLACK_ADMIN_INVITEREQUESTS_APPROVED = ("slack", "admin.inviteRequests.approved")
-    SLACK_APPS_PERMISSIONS_USERS = ("slack", "apps.permissions.users")
-    SLACK_USERGROUPS = ("slack", "usergroups")
-    SLACK_ADMIN_APPS_REQUESTS = ("slack", "admin.apps.requests")
-    SLACK_ADMIN_EMOJI = ("slack", "admin.emoji")
-    SLACK_APPS_PERMISSIONS_RESOURCES = ("slack", "apps.permissions.resources")
-    SLACK_DIALOG = ("slack", "dialog")
-    SLACK_BOTS = ("slack", "bots")
-    SLACK_WORKFLOWS = ("slack", "workflows")
-    SLACK_TEAM = ("slack", "team")
-    SLACK_ADMIN_INVITEREQUESTS_DENIED = ("slack", "admin.inviteRequests.denied")
-    SLACK_ADMIN_USERS = ("slack", "admin.users")
     SLACK_ADMIN_TEAMS_ADMINS = ("slack", "admin.teams.admins")
-    SLACK_MIGRATION = ("slack", "migration")
-    SLACK_FILES = ("slack", "files")
-    SLACK_ADMIN_TEAMS_OWNERS = ("slack", "admin.teams.owners")
-    SLACK_CHAT = ("slack", "chat")
-    SLACK_APPS_EVENT_AUTHORIZATIONS = ("slack", "apps.event.authorizations")
+    SLACK_APPS_PERMISSIONS_SCOPES = ("slack", "apps.permissions.scopes")
+    SLACK_SEARCH = ("slack", "search")
+    SLACKBOT_TEAM_PROFILE = ("slackbot", "team.profile")
+    SLACKBOT_BOTS = ("slackbot", "bots")
+    SLACKBOT_OAUTH_V2 = ("slackbot", "oauth.v2")
+    SLACKBOT_APPS = ("slackbot", "apps")
+    SLACKBOT_APPS_PERMISSIONS = ("slackbot", "apps.permissions")
+    SLACKBOT_CALLS_PARTICIPANTS = ("slackbot", "calls.participants")
+    SLACKBOT_STARS = ("slackbot", "stars")
+    SLACKBOT_WORKFLOWS = ("slackbot", "workflows")
+    SLACKBOT_CALLS = ("slackbot", "calls")
+    SLACKBOT_FILES = ("slackbot", "files")
+    SLACKBOT_REACTIONS = ("slackbot", "reactions")
+    SLACKBOT_APPS_PERMISSIONS_USERS = ("slackbot", "apps.permissions.users")
+    SLACKBOT_APPS_PERMISSIONS_RESOURCES = ("slackbot", "apps.permissions.resources")
+    SLACKBOT_DND = ("slackbot", "dnd")
+    SLACKBOT_EMOJI = ("slackbot", "emoji")
+    SLACKBOT_CONVERSATIONS = ("slackbot", "conversations")
+    SLACKBOT_USERGROUPS = ("slackbot", "usergroups")
+    SLACKBOT_TEAM = ("slackbot", "team")
+    SLACKBOT_FILES_REMOTE = ("slackbot", "files.remote")
+    SLACKBOT_FILES_COMMENTS = ("slackbot", "files.comments")
+    SLACKBOT_DIALOG = ("slackbot", "dialog")
+    SLACKBOT_APPS_EVENT_AUTHORIZATIONS = ("slackbot", "apps.event.authorizations")
+    SLACKBOT_USERS = ("slackbot", "users")
+    SLACKBOT_VIEWS = ("slackbot", "views")
+    SLACKBOT_IMPORTANT = ("slackbot", "important")
+    SLACKBOT_CHAT = ("slackbot", "chat")
+    SLACKBOT_OAUTH = ("slackbot", "oauth")
+    SLACKBOT_AUTH = ("slackbot", "auth")
+    SLACKBOT_USERGROUPS_USERS = ("slackbot", "usergroups.users")
+    SLACKBOT_API = ("slackbot", "api")
     SLACKBOT_REMINDERS = ("slackbot", "reminders")
     SLACKBOT_PINS = ("slackbot", "pins")
-    SLACKBOT_APPS_PERMISSIONS_SCOPES = ("slackbot", "apps.permissions.scopes")
-    SLACKBOT_AUTH = ("slackbot", "auth")
-    SLACKBOT_REACTIONS = ("slackbot", "reactions")
-    SLACKBOT_STARS = ("slackbot", "stars")
-    SLACKBOT_USERGROUPS_USERS = ("slackbot", "usergroups.users")
-    SLACKBOT_OAUTH_V2 = ("slackbot", "oauth.v2")
-    SLACKBOT_FILES_REMOTE = ("slackbot", "files.remote")
-    SLACKBOT_CHAT_SCHEDULEDMESSAGES = ("slackbot", "chat.scheduledMessages")
-    SLACKBOT_DND = ("slackbot", "dnd")
-    SLACKBOT_OAUTH = ("slackbot", "oauth")
-    SLACKBOT_VIEWS = ("slackbot", "views")
-    SLACKBOT_EMOJI = ("slackbot", "emoji")
-    SLACKBOT_USERS_PROFILE = ("slackbot", "users.profile")
-    SLACKBOT_RTM = ("slackbot", "rtm")
-    SLACKBOT_FILES_COMMENTS = ("slackbot", "files.comments")
-    SLACKBOT_CONVERSATIONS = ("slackbot", "conversations")
-    SLACKBOT_APPS_PERMISSIONS = ("slackbot", "apps.permissions")
-    SLACKBOT_API = ("slackbot", "api")
-    SLACKBOT_TEAM_PROFILE = ("slackbot", "team.profile")
-    SLACKBOT_CALLS_PARTICIPANTS = ("slackbot", "calls.participants")
-    SLACKBOT_CALLS = ("slackbot", "calls")
-    SLACKBOT_IMPORTANT = ("slackbot", "important")
-    SLACKBOT_APPS = ("slackbot", "apps")
-    SLACKBOT_APPS_PERMISSIONS_USERS = ("slackbot", "apps.permissions.users")
-    SLACKBOT_USERGROUPS = ("slackbot", "usergroups")
-    SLACKBOT_APPS_PERMISSIONS_RESOURCES = ("slackbot", "apps.permissions.resources")
-    SLACKBOT_DIALOG = ("slackbot", "dialog")
-    SLACKBOT_BOTS = ("slackbot", "bots")
-    SLACKBOT_WORKFLOWS = ("slackbot", "workflows")
-    SLACKBOT_TEAM = ("slackbot", "team")
     SLACKBOT_MIGRATION = ("slackbot", "migration")
-    SLACKBOT_FILES = ("slackbot", "files")
-    SLACKBOT_USERS = ("slackbot", "users")
-    SLACKBOT_CHAT = ("slackbot", "chat")
-    SLACKBOT_APPS_EVENT_AUTHORIZATIONS = ("slackbot", "apps.event.authorizations")
-    SPOTIFY_PLAYLISTS = ("spotify", "Playlists")
+    SLACKBOT_RTM = ("slackbot", "rtm")
+    SLACKBOT_USERS_PROFILE = ("slackbot", "users.profile")
+    SLACKBOT_CHAT_SCHEDULEDMESSAGES = ("slackbot", "chat.scheduledMessages")
+    SLACKBOT_APPS_PERMISSIONS_SCOPES = ("slackbot", "apps.permissions.scopes")
     SPOTIFY_GENRES = ("spotify", "Genres")
-    SPOTIFY_PLAYER = ("spotify", "Player")
-    SPOTIFY_TRACKS = ("spotify", "Tracks")
-    SPOTIFY_ALBUMS = ("spotify", "Albums")
-    SPOTIFY_CHAPTERS = ("spotify", "Chapters")
-    SPOTIFY_USERS = ("spotify", "Users")
     SPOTIFY_MARKETS = ("spotify", "Markets")
-    SPOTIFY_SHOWS = ("spotify", "Shows")
-    SPOTIFY_AUDIOBOOKS = ("spotify", "Audiobooks")
-    SPOTIFY_CATEGORIES = ("spotify", "Categories")
-    SPOTIFY_SEARCH = ("spotify", "Search")
-    SPOTIFY_LIBRARY = ("spotify", "Library")
-    SPOTIFY_EPISODES = ("spotify", "Episodes")
+    SPOTIFY_USERS = ("spotify", "Users")
     SPOTIFY_ARTISTS = ("spotify", "Artists")
+    SPOTIFY_AUDIOBOOKS = ("spotify", "Audiobooks")
+    SPOTIFY_ALBUMS = ("spotify", "Albums")
+    SPOTIFY_TRACKS = ("spotify", "Tracks")
+    SPOTIFY_CHAPTERS = ("spotify", "Chapters")
+    SPOTIFY_SEARCH = ("spotify", "Search")
+    SPOTIFY_PLAYLISTS = ("spotify", "Playlists")
+    SPOTIFY_PLAYER = ("spotify", "Player")
+    SPOTIFY_EPISODES = ("spotify", "Episodes")
+    SPOTIFY_CATEGORIES = ("spotify", "Categories")
+    SPOTIFY_SHOWS = ("spotify", "Shows")
+    SPOTIFY_LIBRARY = ("spotify", "Library")
     TASKADE_FOLDER = ("taskade", "Folder")
-    TASKADE_TASK = ("taskade", "Task")
     TASKADE_ME = ("taskade", "Me")
-    TASKADE_AGENT = ("taskade", "Agent")
+    TASKADE_TASK = ("taskade", "Task")
     TASKADE_WORKSPACE = ("taskade", "Workspace")
     TASKADE_PROJECT = ("taskade", "Project")
-    WHATSAPP_APPLICATION = ("whatsapp", "Application")
-    WHATSAPP_USERS = ("whatsapp", "Users")
-    WHATSAPP_MESSAGES = ("whatsapp", "Messages")
-    WHATSAPP_TWO_STEP_VERIFICATION = ("whatsapp", "Two-Step Verification")
-    WHATSAPP_PROFILE = ("whatsapp", "Profile")
-    WHATSAPP_REGISTRATION = ("whatsapp", "Registration")
+    TASKADE_AGENT = ("taskade", "Agent")
     WHATSAPP_HEALTH = ("whatsapp", "Health")
     WHATSAPP_BUSINESS_PROFILE = ("whatsapp", "Business Profile")
-    WHATSAPP_MEDIA = ("whatsapp", "Media")
+    WHATSAPP_USERS = ("whatsapp", "Users")
     WHATSAPP_CERTIFICATES = ("whatsapp", "Certificates")
     WHATSAPP_CONTACTS = ("whatsapp", "Contacts")
+    WHATSAPP_PROFILE = ("whatsapp", "Profile")
+    WHATSAPP_APPLICATION = ("whatsapp", "Application")
     WHATSAPP_GROUPS = ("whatsapp", "Groups")
     WHATSAPP_BACKUP_RESTORE = ("whatsapp", "Backup/Restore")
-    ZOOM_H323_DEVICES = ("zoom", "H323 Devices")
+    WHATSAPP_REGISTRATION = ("whatsapp", "Registration")
+    WHATSAPP_MEDIA = ("whatsapp", "Media")
+    WHATSAPP_MESSAGES = ("whatsapp", "Messages")
+    WHATSAPP_TWO_STEP_VERIFICATION = ("whatsapp", "Two-Step Verification")
+    ZOOM_WEBINARS = ("zoom", "Webinars")
+    ZOOM_REPORTS = ("zoom", "Reports")
     ZOOM_SIP_PHONE = ("zoom", "SIP Phone")
+    ZOOM_TSP = ("zoom", "TSP")
     ZOOM_ARCHIVING = ("zoom", "Archiving")
+    ZOOM_MEETINGS = ("zoom", "Meetings")
     ZOOM_DEVICES = ("zoom", "Devices")
+    ZOOM_H323_DEVICES = ("zoom", "H323 Devices")
+    ZOOM_CLOUD_RECORDING = ("zoom", "Cloud Recording")
     ZOOM_PAC = ("zoom", "PAC")
     ZOOM_TRACKING_FIELD = ("zoom", "Tracking Field")
-    ZOOM_MEETINGS = ("zoom", "Meetings")
-    ZOOM_TSP = ("zoom", "TSP")
-    ZOOM_CLOUD_RECORDING = ("zoom", "Cloud Recording")
-    ZOOM_REPORTS = ("zoom", "Reports")
-    ZOOM_WEBINARS = ("zoom", "Webinars")
 
 
 
@@ -361,7 +361,7 @@ class App(str, Enum):
     @property
     def is_local(self) -> bool:
         """If the app is local."""
-        return self.value.lower() in ["mathematical", "localworkspace", "cmdmanagertool", "historykeeper", "ragtoolactions", "webtool", "greptile"]
+        return self.value.lower() in ["mathematical", "localworkspace", "cmdmanagertool", "historykeeper", "ragtool", "webtool", "greptile"]
 
     ABLY = "ably"
     ACCELO = "accelo"
@@ -401,7 +401,6 @@ class App(str, Enum):
     BUBBLE = "bubble"
     CAL = "cal"
     CALENDLY = "calendly"
-    CALENDLY1 = "calendly1"
     CHATWORK = "chatwork"
     CHMEETINGS = "chmeetings"
     CLICKUP = "clickup"
@@ -444,17 +443,11 @@ class App(str, Enum):
     GITHUB = "github"
     GITLAB = "gitlab"
     GMAIL = "gmail"
-    GMAIL1 = "gmail1"
-    GOOGLE5 = "google5"
     GOOGLECALENDAR = "googlecalendar"
-    GOOGLECALENDAR3 = "googlecalendar3"
     GOOGLEDOCS = "googledocs"
-    GOOGLEDOCS3 = "googledocs3"
     GOOGLEDRIVE = "googledrive"
-    GOOGLEDRIVE1 = "googledrive1"
     GOOGLEMEET = "googlemeet"
     GOOGLESHEETS = "googlesheets"
-    GOOGLESHEETS1 = "googlesheets1"
     GOOGLETASKS = "googletasks"
     GORGIAS = "gorgias"
     GUMROAD = "gumroad"
@@ -497,7 +490,6 @@ class App(str, Enum):
     NETSUITE = "netsuite"
     NGROK = "ngrok"
     NOTION = "notion"
-    NOTION1 = "notion1"
     OKTA = "okta"
     ONCEHUB = "oncehub"
     ONE_DRIVE = "one-drive"
@@ -577,7 +569,7 @@ class App(str, Enum):
     LOCALWORKSPACE = "localworkspace"
     CMDMANAGERTOOL = "cmdmanagertool"
     HISTORYKEEPER = "historykeeper"
-    RAGTOOLACTIONS = "ragtoolactions"
+    RAGTOOL = "ragtool"
     WEBTOOL = "webtool"
     GREPTILE = "greptile"
 
@@ -826,8 +818,6 @@ class Action(tuple, Enum):
     ASANA_WORK_SPACES_UPDATE_WORK_SPACE_RECORD = ("asana", "asana_work_spaces_update_work_space_record", False)
     ASANA_WORK_SPACES_ADD_USER_TO_WORK_SPACE = ("asana", "asana_work_spaces_add_user_to_work_space", False)
     ASANA_WORK_SPACES_REMOVE_USER_FROM_WORK_SPACE = ("asana", "asana_work_spaces_remove_user_from_work_space", False)
-    ASANA_CREATE_SUBTASK = ("asana", "asana_create_subtask", False)
-    ASANA_GET_SUBTASKS = ("asana", "asana_get_subtasks", False)
     ATTIO_LIST_OBJECTS = ("attio", "attio_list_objects", False)
     ATTIO_CREATE_AN_OBJECT = ("attio", "attio_create_an_object", False)
     ATTIO_GET_AN_OBJECT = ("attio", "attio_get_an_object", False)
@@ -1105,10 +1095,6 @@ class Action(tuple, Enum):
     BREVO_EXTERNAL_FEEDS_GET_FEED_BYU_U_ID = ("brevo", "brevo_external_feeds_get_feed_byu_u_id", False)
     BREVO_EXTERNAL_FEEDS_UPDATE_FEED_BYU_U_ID = ("brevo", "brevo_external_feeds_update_feed_byu_u_id", False)
     BREVO_EXTERNAL_FEEDS_DELETE_FEED_BYU_U_ID = ("brevo", "brevo_external_feeds_delete_feed_byu_u_id", False)
-    CALENDLY1_GET_CALENDLY_USER_INFO = ("calendly1", "calendly1_get_calendly_user_info", False)
-    CALENDLY1_GET_SCHEDULED_EVENTS = ("calendly1", "calendly1_get_scheduled_events", False)
-    CALENDLY1_GET_EVENT_DETAILS = ("calendly1", "calendly1_get_event_details", False)
-    CALENDLY1_CREATE_ONE_OFF_EVENT_TYPE = ("calendly1", "calendly1_create_one_off_event_type", False)
     CLICKUP_ATTACHMENTS_UPLOAD_FILE_TO_TASK_AS_ATTACHMENT = ("clickup", "clickup_attachments_upload_file_to_task_as_attachment", False)
     CLICKUP_AUTHORIZATION_GET_ACCESS_TOKEN = ("clickup", "clickup_authorization_get_access_token", False)
     CLICKUP_AUTHORIZATION_VIEW_ACCOUNT_DETAILS = ("clickup", "clickup_authorization_view_account_details", False)
@@ -2452,17 +2438,6 @@ class Action(tuple, Enum):
     GMAIL_GET_THREAD_DETAILS_WITH_MESSAGES = ("gmail", "gmail_get_thread_details_with_messages", False)
     GMAIL_REPLY_TO_THREAD = ("gmail", "gmail_reply_to_thread", False)
     GMAIL_FETCH_EMAILS_WITH_LABEL = ("gmail", "gmail_fetch_emails_with_label", False)
-    GMAIL1_SEND_EMAIL = ("gmail1", "gmail1_send_email", False)
-    GMAIL1_CREATE_EMAIL_DRAFT = ("gmail1", "gmail1_create_email_draft", False)
-    GMAIL1_FIND_EMAIL_ID = ("gmail1", "gmail1_find_email_id", False)
-    GMAIL1_FETCH_LAST_THREE_MESSAGES = ("gmail1", "gmail1_fetch_last_three_messages", False)
-    GMAIL1_ADD_LABEL_TO_EMAIL = ("gmail1", "gmail1_add_label_to_email", False)
-    GMAIL1_LIST_LABELS = ("gmail1", "gmail1_list_labels", False)
-    GMAIL1_FETCH_MESSAGE_BY_THREAD_ID = ("gmail1", "gmail1_fetch_message_by_thread_id", False)
-    GMAIL1_REPLY_TO_THREAD = ("gmail1", "gmail1_reply_to_thread", False)
-    GMAIL1_FETCH_UNREAD_EMAILS = ("gmail1", "gmail1_fetch_unread_emails", False)
-    GMAIL1_FILTER_EMAILS = ("gmail1", "gmail1_filter_emails", False)
-    GOOGLE5_GET_CONTACTS = ("google5", "google5_get_contacts", False)
     GOOGLECALENDAR_CREATE_GOOGLE_EVENT = ("googlecalendar", "googlecalendar_create_google_event", False)
     GOOGLECALENDAR_REMOVE_ATTENDEE = ("googlecalendar", "googlecalendar_remove_attendee", False)
     GOOGLECALENDAR_FIND_EVENT = ("googlecalendar", "googlecalendar_find_event", False)
@@ -2474,20 +2449,9 @@ class Action(tuple, Enum):
     GOOGLECALENDAR_PATCH_GOOGLE_CALENDAR = ("googlecalendar", "googlecalendar_patch_google_calendar", False)
     GOOGLECALENDAR_QUICK_ADD_GOOGLE_CALENDAR = ("googlecalendar", "googlecalendar_quick_add_google_calendar", False)
     GOOGLECALENDAR_GET_CURRENT_DATE_TIME = ("googlecalendar", "googlecalendar_get_current_date_time", False)
-    GOOGLECALENDAR3_CREATE_GOOGLE_EVENT = ("googlecalendar3", "googlecalendar3_create_google_event", False)
-    GOOGLECALENDAR3_REMOVE_ATTENDEE = ("googlecalendar3", "googlecalendar3_remove_attendee", False)
-    GOOGLECALENDAR3_FIND_EVENT = ("googlecalendar3", "googlecalendar3_find_event", False)
-    GOOGLECALENDAR3_CHECK_ATTENDEES = ("googlecalendar3", "googlecalendar3_check_attendees", False)
-    GOOGLECALENDAR3_DELETE_GOOGLE_EVENT = ("googlecalendar3", "googlecalendar3_delete_google_event", False)
-    GOOGLECALENDAR3_UPDATE_GOOGLE_EVENT = ("googlecalendar3", "googlecalendar3_update_google_event", False)
-    GOOGLECALENDAR3_CREATE_CALENDAR = ("googlecalendar3", "googlecalendar3_create_calendar", False)
     GOOGLEDOCS_UPDATE_EXISTING_DOCUMENT = ("googledocs", "googledocs_update_existing_document", False)
     GOOGLEDOCS_GET_DOCUMENT_BY_ID = ("googledocs", "googledocs_get_document_by_id", False)
     GOOGLEDOCS_CREATE_DOCUMENT = ("googledocs", "googledocs_create_document", False)
-    GOOGLEDOCS3_APPEND_TEXT_TO_DOCUMENT = ("googledocs3", "googledocs3_append_text_to_document", False)
-    GOOGLEDOCS3_CREATE_DOCUMENT_FROM_TEMPLATE = ("googledocs3", "googledocs3_create_document_from_template", False)
-    GOOGLEDOCS3_CREATE_DOCUMENT_FROM_TEXT = ("googledocs3", "googledocs3_create_document_from_text", False)
-    GOOGLEDOCS3_CREATE_DOCUMENT = ("googledocs3", "googledocs3_create_document", False)
     GOOGLEDRIVE_COPY_FILE = ("googledrive", "googledrive_copy_file", False)
     GOOGLEDRIVE_CREATE_FOLDER = ("googledrive", "googledrive_create_folder", False)
     GOOGLEDRIVE_CREATE_FILE_FROM_TEXT = ("googledrive", "googledrive_create_file_from_text", False)
@@ -2497,12 +2461,6 @@ class Action(tuple, Enum):
     GOOGLEDRIVE_EXPORT_FILE = ("googledrive", "googledrive_export_file", False)
     GOOGLEDRIVE_EDIT_FILE = ("googledrive", "googledrive_edit_file", False)
     GOOGLEDRIVE_DELETE_FOLDER_OR_FILE = ("googledrive", "googledrive_delete_folder_or_file", False)
-    GOOGLEDRIVE1_COPY_FILE = ("googledrive1", "googledrive1_copy_file", False)
-    GOOGLEDRIVE1_CREATE_FOLDER = ("googledrive1", "googledrive1_create_folder", False)
-    GOOGLEDRIVE1_CREATE_FILE_FROM_TEXT = ("googledrive1", "googledrive1_create_file_from_text", False)
-    GOOGLEDRIVE1_FIND_FILE = ("googledrive1", "googledrive1_find_file", False)
-    GOOGLEDRIVE1_FIND_FOLDER = ("googledrive1", "googledrive1_find_folder", False)
-    GOOGLEDRIVE1_ADD_FILE_SHARING_PREFERENCE = ("googledrive1", "googledrive1_add_file_sharing_preference", False)
     GOOGLEMEET_GET_MEET = ("googlemeet", "googlemeet_get_meet", False)
     GOOGLEMEET_CREATE_MEET = ("googlemeet", "googlemeet_create_meet", False)
     GOOGLEMEET_GET_CONFERENCE_RECORD_FOR_MEET = ("googlemeet", "googlemeet_get_conference_record_for_meet", False)
@@ -2514,14 +2472,6 @@ class Action(tuple, Enum):
     GOOGLESHEETS_BATCH_UPDATE = ("googlesheets", "googlesheets_batch_update", False)
     GOOGLESHEETS_CLEAR_VALUES = ("googlesheets", "googlesheets_clear_values", False)
     GOOGLESHEETS_BATCH_GET = ("googlesheets", "googlesheets_batch_get", False)
-    GOOGLESHEETS1_CREATE_GOOGLE_SHEET1 = ("googlesheets1", "googlesheets1_create_google_sheet1", False)
-    GOOGLESHEETS1_CREATE_SPREADSHEET_COLUMN = ("googlesheets1", "googlesheets1_create_spreadsheet_column", False)
-    GOOGLESHEETS1_CREATE_SPREADSHEET_ROW = ("googlesheets1", "googlesheets1_create_spreadsheet_row", False)
-    GOOGLESHEETS1_LOOKUP_SPREADSHEET_ROW = ("googlesheets1", "googlesheets1_lookup_spreadsheet_row", False)
-    GOOGLESHEETS1_FIND_WORKSHEET_BY_TITLE = ("googlesheets1", "googlesheets1_find_worksheet_by_title", False)
-    GOOGLESHEETS1_FORMAT_SPREADSHEET_ROW = ("googlesheets1", "googlesheets1_format_spreadsheet_row", False)
-    GOOGLESHEETS1_UPDATE_SPREADSHEET_ROW = ("googlesheets1", "googlesheets1_update_spreadsheet_row", False)
-    GOOGLESHEETS1_FIND_OR_CREATE_WORKSHEET = ("googlesheets1", "googlesheets1_find_or_create_worksheet", False)
     GOOGLETASKS_CREATE_TASK_LIST = ("googletasks", "googletasks_create_task_list", False)
     GOOGLETASKS_DELETE_TASK_LIST = ("googletasks", "googletasks_delete_task_list", False)
     GOOGLETASKS_GET_TASK_LIST = ("googletasks", "googletasks_get_task_list", False)
@@ -2587,9 +2537,6 @@ class Action(tuple, Enum):
     NOTION_INSERT_ROW_DATABASE = ("notion", "notion_insert_row_database", False)
     NOTION_UPDATE_ROW_DATABASE = ("notion", "notion_update_row_database", False)
     NOTION_QUERY_DATABASE = ("notion", "notion_query_database", False)
-    NOTION1_SEARCH_NOTION_PAGE = ("notion1", "notion1_search_notion_page", False)
-    NOTION1_CREATE_NOTION_PAGE = ("notion1", "notion1_create_notion_page", False)
-    NOTION1_CREATE_PAGE_COMMENT = ("notion1", "notion1_create_page_comment", False)
     OKTA_APPLICATION_LIST_APPS = ("okta", "okta_application_list_apps", False)
     OKTA_APPLICATION_CREATE_NEW = ("okta", "okta_application_create_new", False)
     OKTA_APPLICATION_REMOVE_INACTIVE = ("okta", "okta_application_remove_inactive", False)
@@ -3593,7 +3540,7 @@ class Action(tuple, Enum):
     ZOOM_WEB_IN_ARS_GET_WEB_IN_ART_OKEN = ("zoom", "zoom_web_in_ars_get_web_in_art_oken", False)
     ZOOM_WEB_IN_ARS_LIST_TRACKING_SOURCES = ("zoom", "zoom_web_in_ars_list_tracking_sources", False)
     CALCULATOR = ("mathematical", "mathematical_calculator", True, True)
-    WORKSPACESTATUS = ("localworkspace", "localworkspace_workspacestatus", True, True)
+    WORKSPACESTATUSACTION = ("localworkspace", "localworkspace_workspacestatusaction", True, True)
     CREATEWORKSPACEACTION = ("localworkspace", "localworkspace_createworkspaceaction", True, True)
     FINDFILECMD = ("cmdmanagertool", "cmdmanagertool_findfilecmd", True, True)
     CREATEFILECMD = ("cmdmanagertool", "cmdmanagertool_createfilecmd", True, True)
@@ -3607,10 +3554,10 @@ class Action(tuple, Enum):
     GETCURRENTDIRCMD = ("cmdmanagertool", "cmdmanagertool_getcurrentdircmd", True, True)
     GITHUBCLONECMD = ("cmdmanagertool", "cmdmanagertool_githubclonecmd", True, True)
     GETWORKSPACEHISTORY = ("historykeeper", "historykeeper_getworkspacehistory", True, True)
-    RAGTOOLQUERY = ("ragtoolactions", "ragtoolactions_ragtoolquery", True, True)
-    ADDCONTENTTORAGTOOL = ("ragtoolactions", "ragtoolactions_addcontenttoragtool", True, True)
-    SCRAPEWEBSITETOOL = ("webtool", "webtool_scrapewebsitetool", True, True)
-    SCRAPEWEBSITEELEMENTTOOL = ("webtool", "webtool_scrapewebsiteelementtool", True, True)
+    RAGTOOLQUERY = ("ragtool", "ragtool_ragtoolquery", True, True)
+    ADDCONTENTTORAGTOOL = ("ragtool", "ragtool_addcontenttoragtool", True, True)
+    SCRAPEWEBSITECONTENT = ("webtool", "webtool_scrapewebsitecontent", True, True)
+    SCRAPEWEBSITEELEMENT = ("webtool", "webtool_scrapewebsiteelement", True, True)
     CODEQUERY = ("greptile", "greptile_codequery", True, True)
