+++ composio/local_tools/local_workspace/cmd_manager/actions/base_class.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from pydantic import BaseModel, Field
 
-from composio.tools.local import Action
+from composio.core.local import Action
 from composio.local_tools.local_workspace.commons import (
     get_logger,
     HistoryProcessor,
