+++ composio/local_tools/local_workspace/history_keeper/tool.py
@@ -1,6 +1,6 @@
 import typing as t
 from typing import Optional
-from composio.tools.local import Tool, Action
+from composio.core.local import Tool, Action
 
 from composio.local_tools.local_workspace.commons.get_logger import get_logger
 from composio.local_tools.local_workspace.commons import (
