+++ composio/local_tools/ragtool/actions/rag_query.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel, Field
-from composio.tools.local import Action
+from composio.core.local import Action
 
 class RagToolQueryRequest(BaseModel):
     query: str = Field(..., description="The query to search in the knowledge base")
