+++ composio/local_tools/webtool/actions/scrape_website_content.py
@@ -1,7 +1,7 @@
 import ssl
 from pydantic import BaseModel, Field
 from urllib.request import Request, urlopen
-from composio.tools.local import Action
+from composio.core.local import Action
 
 class ScrapeWebsiteToolRequest(BaseModel):
     website_url: str = Field(..., description="Mandatory website url to read contents of the website from")
